data.h 의 존재 여부 확인 후 삭제

+++ PixelBox/Form1.cs
@@ -118,7 +118,7 @@ namespace PixelBox
             p.Start();
             p.WaitForExit();
 
-            File.Delete(".\\rgbmatrix\\data.h");
+            if(File.Exists(".\\rgbmatrix\\data.h")) File.Delete(".\\rgbmatrix\\data.h");
             File.Move(".\\data.h", ".\\rgbmatrix\\data.h");
 
             arduinoUpload.CreateNoWindow = false;
